Move Google Analytics client into ETL folders
This code was leftover from the restructure work in #654. Moving this
with the other Google Analytics classes as a client to GA API.

This is the last thing to move from the services folder and so we
can now get rid of that. Empty folders are automatically removed.
